IO library: implement full 64 bit support with IO buffer and IO buffer data classes in order to support larger than 2GB IO buffers with QFS client. Chunk and meta servers, client and s3 IO libraries: update to work with IO buffer change.

+++ src/cc/libclient/KfsWrite.cc
@@ -161,7 +161,7 @@ KfsClientImpl::Write(int fd, const char *buf, size_t numBytes,
     const KfsProtocolWorker::FileInstance fileInstance = entry.instance;
     const string                          pathName     = entry.pathname;
     const int                             bufsz        = entry.ioBufferSize;
-    const int                             prevPending  = entry.pending;
+    const int64_t                         prevPending  = entry.pending;
     const bool                            throttle     =
         ! asyncFlag && bufsz > 0 && bufsz <= entry.pending;
     if ((throttle || bufsz <= 0) && ! asyncFlag) {
